third_party/WebKit/LayoutTests/http/tests/devtools/startup/sources/debugger/script-formatter-breakpoints-4-expected.txt

Tests the script formatting is working fine with breakpoints.


Running: testSetup

Running: testBreakpointSetInOriginalAndRemovedInFormatted
Adding breakpoint.
Formatting.
Removing breakpoint.
Unformatting.
